+++ test/index.test.js
@@ -39,3 +39,4 @@ const isEmptyObject = require("./cases/isEmptyObject.test.js");
 const isIterable = require("./cases/isIterable.test.js");
 const isPositiveInfinity = require("./cases/isPositiveInfinity.test.js");
 const isNegativeInfinity = require("./cases/isNegativeInfinity.test.js");
+const isArrowFunction = require("./cases/isArrowFunction.test.js");